Center buttons are fixed (#344)
* fixed center view resume dialog

* fixed center view

fixed newgamesettings2 for having int values in bigblind etc.
game options button is hidden for game ended screen.

* Used double for blinds

* Change numeric keyboard colors and sizes

+++ lib/models/game/new_game_model.dart
@@ -63,7 +63,7 @@ class NewGameModel {
   int gameLength = 60;
   bool buyInApproval = true;
   double rakePercentage = 0;
-  int rakeCap = 0;
+  double rakeCap = 0;
   int buyInMin = 30;
   int buyInMax = 100;
   int actionTime = 30;
@@ -197,10 +197,10 @@ class NewGameModel {
 }
 
 class Blinds {
-  double smallBlind = 1.0;
-  double bigBlind = 2.0;
-  double straddle = 4.0;
-  double ante = 0.0;
+  double smallBlind = 1;
+  double bigBlind = 2;
+  double straddle = 4;
+  double ante = 0;
 
   Blinds({this.smallBlind, this.bigBlind, this.straddle, this.ante});
 }
